+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 0.1.2 Under development
 
+- Bug #34: Add `phpstan` type hints for `StatelessApplication` in `FrankenPHP` and `TestCase` classes (@terabytesoftw)
+
 ## 0.1.1 January 25, 2026
 
 - Bug #30: Update `tests/support/bootstrap.php` file path and add `TestCase` class for improved testing structure (@terabytesoftw)